New message showed to the user when it is not possible to establish a database connection.
Catched errors caused by Tomcat security enable and logged to error log, to help the user detect the problem. In some situations it is possible to show a message to the final user, but in others it is not possible like:

        - "blocked access to DNS": the program just issues a line to the log
file because I think it is not convenient to stop the program just
because we can't show hostnames.

        - If Tomcat security is enable jfreechart can't create an image from a chart. This error is logged to the error log but it is not shown to the user because this error happens in the Servlet that creates the image
and which is called as a image then can't return a text, unless we
create an image with the error message. At the moment we are just
issuing a message to the log file.

git-svn-id: https://svn.code.sf.net/p/pmgraph/code/pmgraph@160 903098bd-d3be-4b5e-9df1-ec1ba492e1ff